move text or url to types
